Move AbstractSnapshotsContainer to ContainerTraits.hpp
